Move shared-session guard into middleware
